requirements.txt
asgiref==3.4.1
cffi==1.15.0
cryptography==36.0.1
Django==4.0
django-ninja==0.16.1
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
idna==3.3
Pillow==8.4.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.8.2
python-jose==3.3.0
rsa==4.8
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.0.1
tzdata==2021.5